Fix language major mode editing
When this project was renamed recently from `picolisp-mode` to
`plisp-mode`, language major mode editing a la:

    https://orgmode.org/manual/Editing-source-code.html

stopped working, that is, picolisp source code blocks in Org Babel
documents could not be edited in language major mode when the user
pressed C-'.

This fix allows language major mode editing to work once again.

Background. When the user presses C-' while point is in a picolisp
source code block, Org Babel is looking for the major mode
`picolisp-mode` (*exactly*) and fails to find it, absent this patch.
The patch simply provides the feature and the entry function
`picolisp-mode`.